apps/llmphysics-bot/praw-bot/requirements.txt
praw>=7.7
wikipedia-api>=0.6
python-dotenv>=1.0
apscheduler>=3.10
